backend/requirements.txt
fastapi>=0.115.0,<0.120.0
uvicorn>=0.30.0,<0.40.0
motor>=3.5.0,<4.0.0
pydantic>=2.10.0,<3.0.0
python-jose[cryptography]>=3.5.0,<4.0.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0,<5.1.0
python-multipart>=0.0.15,<1.0.0
email-validator>=2.0.0,<3.0.0
pymongo>=4.10.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0